+++ integration_tests/suite/helpers/fixtures.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright 2017 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2018 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0+
 
 import os
@@ -202,7 +202,13 @@ def policy(**policy_args):
 
 def tenant(**tenant_args):
     if 'name' not in tenant_args:
-        tenant_args['name'] = _random_string(20)
+        tenant_args['name'] = None
+    if 'phone' not in tenant_args:
+        tenant_args['phone'] = None
+    if 'contact' not in tenant_args:
+        tenant_args['contact'] = None
+    if 'address_id' not in tenant_args:
+        tenant_args['address_id'] = None
 
     def decorator(decorated):
         @wraps(decorated)